% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabomapr.R
\name{adjacency_edgeList}
\alias{adjacency_edgeList}
\title{CID_tanimoto}
\usage{
adjacency_edgeList(mat, symmetric = TRUE, diagonal = FALSE)
}
\arguments{
\item{mat}{adjacency matrix}

\item{symmetric}{TRUE if undirected keep lower.tri}

\item{diagonal}{TRUE to keep self-connections}

\item{mat}{adjacency matrix}
}
\details{
convert adjacency matrix into an edge list
}

