% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxp.R
\name{boxp}
\alias{boxp}
\title{box}
\usage{
boxp(e, p, color, selected = NULL)
}
\arguments{
\item{e}{should be the e, where rows are samples and columns are compounds.}
}
\description{
box
}
\details{
https://plot.ly/~jackp/969/points-scored-by-the-top-50-scoring-nba-players-in-2012-each-point-is-a-game/#code
}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

