% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callBionet.R
\name{callBionet}
\alias{callBionet}
\title{Compute a maximum scoring subnetwork}
\usage{
callBionet(edgelist, nodelist, pval, fdr)
}
\arguments{
\item{edgelist}{a data frame of edges}

\item{nodelist}{a data frame of nodes containin node information e.g. node name, node xref. Default is NULL.}

\item{pval}{a numeric vector of p-values with name attributes}

\item{fdr}{a numeric value specifying false discovery rate. Default is 0.05.}
}
\value{
list of nodes and edges. The list contains the data frame of nodes and the data frame of edges.
Return empty list of data frame if error or found nothing.
}
\description{
compute a maximum scoring subnetwork or module by integrating omic data and the biological network.
The function wraps around the main functions of \pkg{\link{BioNet}} including \code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}.
}
\details{
The method \code{bionet} encapsulates the main functions from \pkg{\link{BioNet}} to compute a functional module or significanly changed subnetworks.
Those functions include: \code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}.
}
\examples{
#kw <- c('G15729','G17561','G16015','G18145','G16708')
#grinnNW <- fetchGrinnNetwork(txtInput=kw, from="metabolite", to="protein")
#library(grinn)
#data(dummyStat)
#result <- fetchSubnetwork(statInput = dummyStat, nwInput = grinnNW)
#library(igraph)
#plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Beisser D., Klau GW., Dandekar T., Müller T. and Dittrich MT. (2010) BioNet: an R-Package for the functional analysis of biological networks. Bioinformatics, 26(8):1129-30

Dittrich MT., Klau GW., Rosenwald A., Dandekar T., Müller T. (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. Bioinformatics, 24(13):i223-31
}
\seealso{
\code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}
}

