% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMesh.R
\name{callMesh}
\alias{callMesh}
\title{Retrieve Mesh from PubChem}
\usage{
callMesh(pcid)
}
\arguments{
\item{pcid}{a string of PubChem CID}
}
\value{
list of data frame of nodes and data frame of edges with the following components:

nodes:

\code{id} = mesh id

\code{gid} = mesh id

\code{nodename} = mesh name

\code{nodelabel} = Mesh

#'\code{nodexref} = mesh id

edges:

\code{source, target} = node

\code{type} = ANNOTATION

\code{datasource} = PUBCHEM

Return empty list of data frame if error or found nothing.
}
\description{
retrieve Mesh terms for a given PubChem CID from PubChem
}
\examples{
# mesh = callMesh(345)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}

