% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWordCloud.R
\name{callWordCloud}
\alias{callWordCloud}
\title{Compute wordCloud}
\usage{
callWordCloud(edgelist, nodelist)
}
\arguments{
\item{edgelist}{a data frame of annotation pairs. A source column (1st column) contains annotation terms and a target column (2nd column) contains annotated entities.}

\item{nodelist}{a data frame of node attributes e.g. node neo4j id, node grinn id, node name, node xref.}
}
\value{
data frame of wordcloud contains the following components:

\code{rank} = rank sort by freq

\code{id} = annotation id or annotation neo4j id

\code{gid} = annotation id or annotation grinn id

\code{nodename} = annotation name

\code{nodelabel} = annotation type

\code{nodexref} = cross references

\code{freq} = frequency of the annotation term

\code{member} = list of members of the annotation term

Return empty data frame if error or found nothing.
}
\description{
compute wordCloud of given annotation pairs retrieved from e.g. \code{\link{fetchNetwork}}.
}
\examples{
#simnw <- computeSimilarity(c(1110,10413,196,51,311,43,764,790)) #compute similarity network for given pubchem compounds
#result <- computeNwWordCloud(simnw$edges, simnw$nodes, annotation="mesh", internalid=FALSE)
#wordcloud::wordcloud(words = result$wordcloud$nodename, freq = result$wordcloud$freq, scale=c(2,.1),min.freq = 1,max.words=50, random.order=FALSE, rot.per=0.5, colors=RColorBrewer::brewer.pal(8, "Dark2"))
#barplot(result$wordcloud$freq[1:10], las = 2, names.arg = result$wordcloud$nodename[1:10], col ="lightblue", main ="Most frequent words", ylab = "Word frequencies")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
http://www.sthda.com/english/wiki/text-mining-and-word-cloud-fundamentals-in-r-5-simple-steps-you-should-know
}
\seealso{
\pkg{\link{tm}}, \pkg{\link{wordcloud}}

\pkg{\link{tm}}
}

