% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_get_packages.R
\name{check_get_packages}
\alias{check_get_packages}
\title{Check libraries needed.}
\usage{
check_get_packages(pkg)
}
\arguments{
\item{pkg}{a vector containing the names(string) of packages needed.}
}
\description{
Check and download libraries needed.
Check if library already downloaded then either load it or download it and then load it.
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

