% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelation.R
\name{computeCorrelation}
\alias{computeCorrelation}
\title{Compute weighted correlation network}
\usage{
computeCorrelation(x, y, xtype, ytype, internalid, coef, pval, method, returnas)
}
\arguments{
\item{x}{a data frame of raw or quantified data e.g. gene expression data, metabolite intensities.
Columns are samples and rows are molecular entities e.g. genes, proteins or compounds.}

\item{y}{a data frame of raw or quantified data e.g. gene expression data, metabolite intensities.
Columns are samples and rows are molecular entities e.g. genes, proteins or compounds. See details.}

\item{xtype}{a string specifying a node type (default = NULL). If provided and the database is installed, node attributes will be automatically retrieved from the database.
For database query, the value can be one of compound, protein, gene, rna, dna.}

\item{ytype}{a string specifying a node type (default = NULL).}

\item{internalid}{a logical value indicating whether the network nodes are neo4j ids, if TRUE (default). See \code{\link{convertId}} for how to convert ids.
It has no effect if \code{xtype} = NULL or there is no database installed.}

\item{coef}{a numeric value specifying the minimum absolute correlation coefficient to be included in the output (from 0 to 1, default is 0.7).}

\item{pval}{a numeric value specifying the maximum p-value to be included in the output (default is 0.05).}

\item{method}{a string specifying method for computing correlation. It can be one of pearson, kendall, spearman (default). See \code{\link{cor}} for details.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of network information with the following components:

nodes:

\code{id} = node id or node neo4j id

\code{gid} = node id or node grinn id

\code{nodename} =  node id or node name

\code{nodelabel} = node type if provided

edges:

\code{source, target} = node id or node neo4j id

\code{coef} = correlation coefficient

\code{pval} = p-value

\code{direction} = correlation direction

\code{type} = relationship type

Return empty list if error or found nothing.
}
\description{
compute the correlation network of entities from one or two quantified data sets, see details.
}
\details{
The function wraps around \code{\link{corAndPvalue}} function of \pkg{\link{WGCNA}}. Correlation coefficients, p-values and correlation directions are calculated.
The correlation coefficients are continuous values between -1 (negative correlation) and 1 (positive correlation), with numbers close to 1 or -1, meaning very closely correlated.

x and y are data frame in which columns are samples and rows are entities.
If y is given, then the correlations between the x entities and y entities are computed. Otherwise the correlations between x entities are computed.
}
\note{
If only one data set \code{x} is provided and the database is installed, node attributes will be automatically retrieved from the database. Otherwise node attributes will be the original input.
}
\examples{
# Compute a correlation network from compound data
#dt <- data.frame(id=seq(32), mtcars, row.names = NULL) #data frame of x
#result <- computeCorrelation(x = dt, xtype = "compound", coef=0.9, pval = 1e-10)
# Compute a correlation network from two data sets
#dtX <- data.frame(id=seq(16), mtcars[1:16,], row.names = NULL) #data frame of x
#dtY <- data.frame(id=seq(17,32), mtcars[17:32,], row.names = NULL) #data frame of x
#result <- computeCorrelation(x=dtX, y=dtY, coef=0.9, pval = 1e-7)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Langfelder P. and Horvath S. (2008) WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics, 9:559

Dudoit S., Yang YH., Callow MJ. and Speed TP. (2002) Statistical methods for identifying differentially expressed genes in replicated cDNA microarray experiments, STATISTICA SINICA, 12:111

Langfelder P. and Horvath S. Tutorials for the WGCNA package \url{http://labs.genetics.ucla.edu/horvath/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/index.html}
}
\seealso{
\code{\link{cor}}, \code{\link{corAndPvalue}}
}

