% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnrichment.R
\name{computeEnrichment}
\alias{computeEnrichment}
\title{Perform enrichment analysis}
\usage{
computeEnrichment(edgelist, pval, fc, method, size, returnas)
}
\arguments{
\item{edgelist}{a two-column data frame of molecular entities and entity set collections. 
The first column contains molecular entities and the second column contains entity set collections. See \code{\link{loadGSC}} for details.}

\item{pval}{a numeric vector containing the p-values of molecular entities, computed from statistical analysis, 
with name attributes corresponding to the entities. See \code{\link{runGSA}} for details.}

\item{fc}{a numeric vector containing fold changes or sign information (positive or negative) of molecular entities 
with name attributes corresponding to the entities. See \code{\link{runGSA}} for details.}

\item{method}{a string specifying the enrichment analysis method. It can be one of reporter (default), fisher, median, mean, stouffer. See \code{\link{runGSA}}}

\item{size}{a numeric vector specifying the minimum and maximum number of members in each entity set collection to be used in the analysis. Default is c(3,500).}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
entity sets with the following components:

\code{id} = id of entity set

\code{amount} = number of members in each entity set

\code{Stat} = entity set statistics

\code{p} = entity set p-values

\code{p adj} = adjusted entity set p-values

\code{member} = list of entities
See \code{\link{runGSA}} for more details about directional classes. Return empty list or data frame if error or found nothing.
}
\description{
perform enrichment analysis by integrating the statistical values of molecular entities and the compound, protein or gene set collections.
The function wraps around the main functions of \pkg{\link{piano}}.
}
\examples{
#kw <- c('G15729','G17561','G16015','G18145','G16708')
#grinnNW <- fetchGrinnNetwork(txtInput=kw, from="metabolite", to="protein")
#library(grinn)
#data(dummyStat)
#result <- fetchSubnetwork(statInput = dummyStat, nwInput = grinnNW)
#library(igraph)
#plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Fisher R. (1932) Statistical methods for research workers. Oliver and Boyd, Edinburgh.

Stouffer S., Suchman E., Devinney L., Star S., and Williams R. (1949) The American soldier: adjustment during army life. Princeton University Press, Oxford, England.

Patil K. and Nielsen J. (2005) Uncovering transcriptional regulation of metabolism by using metabolic network topology. Proceedings of the National Academy of Sciences of the United States of America 102(8), 2685.

Oliveira A., Patil K., and Nielsen J. (2008) Architecture of transcriptional regulatory circuits is knitted over the topology of bio-molecular interaction networks. BMC Systems Biology 2(1), 17.

Väremo L., Nielsen J., and Nookaew I. (2013) Enriching the gene set analysis of genome-wide data by incorporating directionality of gene expression and combining statistical hypotheses and methods. Nucleic Acids Research, 41(8), pp. 4378-4391.
}
\seealso{
\code{\link{loadGSC}}, \code{\link{runGSA}}, \code{\link{GSAsummaryTable}}
}

