% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNodeEnrichment.R
\name{computeNodeEnrichment}
\alias{computeNodeEnrichment}
\title{Perform enrichment analysis on the input entities}
\usage{
computeNodeEnrichment(nodedata, nodetype, annotation, internalid, method, size, returnas)
}
\arguments{
\item{nodedata}{a data frame of entities and the statistical values. #'The first column is id and the second is e.g. p-values.}

\item{nodetype}{a string specifying the type of source nodes. It can be one of compound, protein, gene, rna, dna.}

\item{annotation}{a string specifying the type of annotations e.g. pathway and mesh. Mesh annotation is available for PubChem compounds only.}

\item{internalid}{boolean value, whether nodedata use neo4j ids, see \code{\link{convertId}} for how to convert to neo4j ids.
\code{internalid} has no effect on Mesh enrichment.}

\item{method}{a string specifying the enrichment analysis method. It can be one of reporter (default), fisher, median, mean, stouffer. See \code{\link{runGSA}}}

\item{size}{a numeric vector specifying the minimum and maximum number of members in each entity set collection to be used in the analysis. Default is c(3,500).}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of nodes, edges and enrichment analysis result. The list contains the data frame of nodes, the data frame of annotation pairs and
the data frame of enrichment analysis result with the following components:

\code{rank} = rank sort by p adj

\code{id} = internal neo4j id

\code{gid} = grinn id

\code{nodename} = name of entity set

\code{nodelabel} = annotation type

\code{nodexref} = cross references

\code{p} = entity set p-values

\code{p adj} = adjusted entity set p-values

\code{no_of_entities} = number of input entities in each entity set

\code{total} = total number of entities in each entity set from the database, not available for Mesh

\code{member} = list of entity members of the entity set
See \code{\link{runGSA}} for more details about directional classes. Return empty list if error or found nothing.
}
\description{
perform enrichment analysis on the list of entities (e.g. compound, protein, gene) by integrating with the statistical values.
The function wraps around \code{\link{computeEnrichment}} for enrichment analysis.
}
\examples{
#kw <- c('G15729','G17561','G16015','G18145','G16708')
#grinnNW <- fetchGrinnNetwork(txtInput=kw, from="metabolite", to="protein")
#library(grinn)
#data(dummyStat)
#result <- fetchSubnetwork(statInput = dummyStat, nwInput = grinnNW)
#library(igraph)
#plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Fisher R. (1932) Statistical methods for research workers. Oliver and Boyd, Edinburgh.

Stouffer S., Suchman E., Devinney L., Star S., and Williams R. (1949) The American soldier: adjustment during army life. Princeton University Press, Oxford, England.

Patil K. and Nielsen J. (2005) Uncovering transcriptional regulation of metabolism by using metabolic network topology. Proceedings of the National Academy of Sciences of the United States of America 102(8), 2685.

Oliveira A., Patil K., and Nielsen J. (2008) Architecture of transcriptional regulatory circuits is knitted over the topology of bio-molecular interaction networks. BMC Systems Biology 2(1), 17.

Väremo L., Nielsen J., and Nookaew I. (2013) Enriching the gene set analysis of genome-wide data by incorporating directionality of gene expression and combining statistical hypotheses and methods. Nucleic Acids Research, 41(8), pp. 4378-4391.
}
\seealso{
\code{\link{loadGSC}}, \code{\link{runGSA}}, \code{\link{GSAsummaryTable}}
}

