% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNodeWordCloud.R
\name{computeNodeWordCloud}
\alias{computeNodeWordCloud}
\title{Compute WordCloud the input entities}
\usage{
computeNodeWordCloud(txtinput, nodetype, annotation, internalid, returnas)
}
\arguments{
\item{txtinput}{a character vector of entities used for the query e.g. txtinput = c('id1', 'id2').
The value can be neo4j ids or grinn ids, see \code{\link{convertId}} for how to convert ids.}

\item{nodetype}{a string specifying the type of source nodes. It can be one of compound, protein, gene, rna, dna.}

\item{annotation}{a string specifying the type of annotations e.g. pathway and mesh. Mesh annotation is available for PubChem compounds only.}

\item{internalid}{boolean value, whether txtinput are neo4j ids, see \code{\link{convertId}} for how to convert ids.
\code{internalid} has no effect on Mesh enrichment.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of nodes, edges and WordCloud result. The list contains the data frame of nodes, the data frame of annotation pairs and
the data frame of wordCloud result with the following components:

\code{rank} = rank sort by freq

\code{id} = internal neo4j id

\code{gid} = grinn id

\code{nodename} = name of entity set

\code{nodelabel} = annotation type

\code{nodexref} = cross references

\code{freq} = frequency of words

\code{member} = list of members of the entity set
}
\description{
compute WordCloud for the list of entities (e.g. compound, protein, gene).
The function wraps around \pkg{\link{tm}}.
}
\examples{
#kw <- c('G15729','G17561','G16015','G18145','G16708')
#grinnNW <- fetchGrinnNetwork(txtInput=kw, from="metabolite", to="protein")
#library(grinn)
#data(dummyStat)
#result <- fetchSubnetwork(statInput = dummyStat, nwInput = grinnNW)
#library(igraph)
#plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
http://www.sthda.com/english/wiki/text-mining-and-word-cloud-fundamentals-in-r-5-simple-steps-you-should-know
}
\seealso{
\pkg{\link{tm}}, \pkg{\link{wordcloud}}
}

