% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{computeSimilarity}
\alias{computeSimilarity}
\title{Compute Tanimoto similarity network}
\usage{
computeSimilarity(txtinput, coef, returnas)
}
\arguments{
\item{txtinput}{a character vector of PubChem CIDs.}

\item{coef}{a numeric value specifying the minimum Tanimoto similarity correlation coefficient to be included in the output (from 0 to 1, default is 0.7).}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of network information with the following components:

nodes:

\code{id} = pubchem CID or node neo4j id

\code{gid} = pubchem CID

\code{nodename} = pubchem CID or node name

\code{nodelabel} = node type

edges:

\code{source, target} = pubchem CID or node neo4j id

\code{coef} = Tanimoto similarity coefficient

\code{type} = relationship type

Return empty list if error or found nothing.
}
\description{
compute a network for compounds based on chemical structure similarity.
The function computes Tanimoto distances between the given compounds using PubChem fingerprints.
}
\note{
If the database is installed, node attributes will be automatically retrieved from the database. Otherwise node attributes will be the original input.
}
\examples{
#simnw <- computeSimilarity(c(1110,10413,196,51,311,43,764,790)) #compute similarity network for given pubchem compounds
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Willett P., Barnard JM. and Downs GM. (1998) Chemical similarity searching. J. Chem. Inf. Comput. Sci., 38, 983–996.

Barupal KD., et al. (2012) MetaMapp: mapping and visualizing metabolomic data by integrating information from biochemical pathways and chemical and mass spectral similarity. BMC Bioinformatics. 13:99.

Grapov D., Wanichthanarak K. and Fiehn O. (2015) MetaMapR: pathway independent metabolomic network analysis incorporating unknowns. Bioinformatics. 31(16):2757-60.

Cao Y., Charisi A., Cheng L., Jiang T. and Girke T. (2008) ChemmineR: a compound mining framework for R. Bioinformatics, 24(15), pp. 1733–1734.

\url{ftp://ftp.ncbi.nih.gov/pubchem/specifications/pubchem_fingerprints.txt}
}
\seealso{
\pkg{\link{metabomapr}}, \code{\link{CID_tanimoto}}, \code{\link{fpSim}}, \code{\link{sdfid}}
}

