% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSubnetwork.R
\name{computeSubnetwork}
\alias{computeSubnetwork}
\title{Compute a subnetwork}
\usage{
computeSubnetwork(edgelist, nodelist, pval, fc, fdr, internalid, method, returnas)
}
\arguments{
\item{edgelist}{a data frame of edges contains at least a source column (1st column) and a target column (2nd column).}

\item{nodelist}{a data frame of nodes contains node attributes e.g. node id, node name, node xref. Default is NULL.}

\item{pval}{a numeric vector or a two-column data frame of statistical values e.g. p-values.
If \code{pval} is a vector, the name attributes must be identical to the names of network nodes.
If \code{pval} is a data frame, 1st column contains the network nodes and 2nd column contains statistical values.}

\item{fc}{a numeric vector of fold changes with name attributes identical to the names of network nodes. Default is NULL.}

\item{fdr}{a numeric value specifying false discovery rate. Default is 0.05.}

\item{internalid}{a logical value indicating whether the network nodes are neo4j ids, if TRUE (default).
It has no effect if there is no database installed.}

\item{method}{a string specifying the method used to compute the subnetwork. Default is bionet.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of network information with the following components:

nodes:

\code{nodeAttributes} = node attributes provided

\code{score} = node score if the method is \code{bionet}: positive values = signal content and negative values = background noise

edges:

\code{source, target}

\code{edgeAttributes} = edge attributes provided

Return empty list if error or found nothing.
}
\description{
compute a subnetwork from p-values of the nodes of a network.
}
\details{
For the method \code{bionet}, the function wraps around the main steps of \pkg{\link{BioNet}} including \code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}
to compute a subnetwork.
}
\examples{
#simnw <- computeSimilarity(c(1110,10413,196,51,311,43,764,790)) #compute similarity network for given pubchem compounds
#pval <- data.frame(pubchem=c(1110,10413,196,51,311,43,764,790), stat=runif(8, 0, 0.06)) #statistical values of pubchem compounds
#result <- computeSubnetwork(simnw$edges, simnw$nodes, pval=pval, internalid = F)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Beisser D., Klau GW., Dandekar T., Müller T. and Dittrich MT. (2010) BioNet: an R-Package for the functional analysis of biological networks. Bioinformatics, 26(8):1129-30

Dittrich MT., Klau GW., Rosenwald A., Dandekar T., Müller T. (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. Bioinformatics, 24(13):i223-31
}
\seealso{
\code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}
}

