% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMeshTree.R
\name{createMeshTree}
\alias{createMeshTree}
\title{Create Mesh tree}
\usage{
createMeshTree(meshtable,fntype)
}
\arguments{
\item{meshtable}{data frame of mesh results or list of data frame of nodes, edges, mesh results.
The mesh results contain with the following columns: id, Tree, MeshName, member, p_adj (for enrichment and overrepresentation)or freq (for wordcloud).}

\item{fntype}{a string specifying a method calculating mesh results. It can be one of wordcloud, overrep, enrichment.}

\item{toprank}{a numeric number specifying the top rank of mesh results. Default is 50.}
}
\value{
list of mesh tree from \code{\pkg{data.tree}}
}
\description{
format mesh results into \code{\pkg{data.tree}} structure. Mesh results are from enrichment, overrepresentation and worldcloud analysis.
}
\details{
\code{\pkg{data.tree}} is used to convert mesh results to mesh tree structure. The tree can be plotted by d3js on GUI.
}
\examples{
#txtinput <- c(1110,10413,196,51,311,43,764,790) #compute overrepresented terms for given pubchem compounds
#nodeoverr <- computeNodeOverrep(txtinput=txtinput, nodetype="compound", annotation="mesh", internalid=FALSE)
nodeoverr = merge(nodeoverr$overrepresentation,MESH,by.x='id',by.y='MeshId')
#result = createMeshTree(nodeoverr, fntype="overrep")
#jsonlite::toJSON(result)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\pkg{data.tree}}, \url{https://d3js.org/}
}

