% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat rm_outlier.R
\name{Deal_with_outlier}
\alias{Deal_with_outlier}
\title{Remove Outlier}
\usage{
Deal_with_outlier(data, method = "remove")
}
\description{
Remove Outlier
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

