% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchHetNetworkByGID.R
\name{fetchHetNetworkByGID}
\alias{fetchHetNetworkByGID}
\title{Query networks containing one or several relationship types from the database}
\usage{
fetchHetNetworkByGID(from, to, pattern, returnas)
}
\arguments{
\item{from}{a character vector of start nodes e.g. from = c('id1', 'id2'). Given \code{from} = NULL, will result in all possible start nodes.
Otherwise the value must be a grinn id, see details and see \code{\link{convertId}} for how to convert ids.}

\item{to}{a character vector of end nodes e.g. to = c('id1', 'id2'). Given \code{to} = NULL, will result in all possible end nodes, see \code{from} for details.}

\item{pattern}{a string specifying the relationship pattern.

Providing a node as \code{(from:Nodetype)} is to specify the start node.
Providing a node as \code{(to:Nodetype)} is to specify the end node.
Providing a node as \code{(:Nodetype)} means any node of the given node type.

A relationship type is indicated by a pair of square brackets between the arrow e.g. \code{-[:RELATIONSHIP_TYPE]->}}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of network information with the following components:

nodes:

\code{id} = node neo4j id

\code{gid} = node grinn id

\code{nodename} = node name

\code{nodelabel} = node type

\code{nodexref} = node cross references

edges:

\code{source, target} = node neo4j id

\code{type} = relationship type

\code{datasource} = relationship resource

\code{properties} = relationship properties

Return empty list if error or found nothing.
}
\description{
query networks of a relationship pattern that contains the given \code{from} and/or \code{to} nodes using grinn id (gid), see details.
}
\details{
The function is used to query networks containing one or more relationship types (heterogeneous network). Use \code{\link{fetchNetworkByGID}} to query networks containing one relationship type.

The database uses two id systems. The neo4j id is a numeric, internal id automatically generated by the database system.
The grinn id (gid) is an id system of Grinn database that uses main ids of standard resources
i.e. ENSEMBL for genes (e.g.ENSG00000139618), UniProt for proteins (e.g.P0C9J6), PubChem CID for compounds (e.g.5793), KEGG for pathways (e.g.hsa00010).
}
\examples{
# Query the network of (from:Compound)-BIOCHEMICAL_REACTION->(to:Compound)<-ANNOTATION-(:Pathway)
#from = list('1060','284','760') #By default, PubChem ids can be used as grinn ids for compounds
#to = list('222656','107689','5950','71080') #By default, PubChem ids can be used as grinn ids for compounds
#pattern = "(from:Compound)-[:BIOCHEMICAL_REACTION]->(to:Compound)<-[:ANNOTATION]-(:Pathway)"
#result = fetchHetNetworkByGID(from=from, to=to, pattern=pattern)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{convertId}}, \code{\link{fetchNetworkByGID}}

For database structure see \url{http://grinnhomepage}
}

