% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchNode.R
\name{fetchNode}
\alias{fetchNode}
\title{Query nodes from the database}
\usage{
fetchNode(txtinput, nodetype, searchby, exactmatch, returnas)
}
\arguments{
\item{txtinput}{a character vector of entities.
It can be the value of any node property key e.g. \code{txtinput} = c('pubchemId1', 'pubchemId2'), see also \code{searchby}.}

\item{nodetype}{a string specifying a node type. It can be one of compound, protein, gene, pathway, rna, dna, phenotype.}

\item{searchby}{a string specifying a node property key used for the query. It can be one of neo4jid (default), grinnid, name, synonyms, description, properties, xref, datasource.}

\item{exactmatch}{a logical value. If TRUE (default), match exactly, case-sensitive and data type-sensitive.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
node information. Return empty list or data frame if error or found nothing.
}
\description{
query nodes from the database which will return node information or node attributes
including neo4jid, grinnid, nodename, nodelabel, datasource, description, organism, synonyms and xref, if applicable.
}
\details{
The database uses two id systems. The neo4j id is a numeric, internal id automatically generated by the database system.
The grinn id (gid) is an id system of Grinn database that uses main ids of standard resources
i.e. ENSEMBL for genes (e.g.ENSG00000139618), UniProt for proteins (e.g.P0C9J6), PubChem CID for compounds (e.g.5793), KEGG for pathways (e.g.hsa00010).
}
\note{
maximum no. of nodes returned = 30000 nodes
}
\examples{
# Query compounds by name
#kw <- list('1-Methylhistidine','D-Lactic acid')
#result <- fetchNode(txtinput=kw, nodetype="compound", searchby="name")
# Query compounds by KEGG ids
#kw <- c('C01152','C00256','C00345')
#result <- fetchNode(txtinput=kw, nodetype="compound", searchby="xref")
# Query compounds by synonyms
#kw <- '1-Methyl histidine'
#result <- fetchNode(txtinput=kw, nodetype="compound", searchby="synonyms", exactmatch=FALSE)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}

