% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNetworkOutput.R
\name{formatNetworkOutput}
\alias{formatNetworkOutput}
\title{Format resulting network}
\usage{
formatNetworkOutput(graph, returnas)
}
\arguments{
\item{graph}{graph output from neo4j to be formatted}
}
\value{
relationship information with the following components:
\code{source, target} = node internal id

\code{type} = relationship type

\code{datasource} = relationship resource

\code{properties} = relationship properties

Return empty list or data frame if error or found nothing.
}
\description{
format graph output from neo4j to data frame of network.
}
\examples{
#cypher = "MATCH ptw = (from:Protein)-[:CONTROL]->(to:Gene) WHERE from.GID = 'P14859' RETURN DISTINCT ptw LIMIT 2"
#nw = curlRequest.TRANSACTION(cypher=qstring)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{fetchNetwork}}, \code{\link{fetchNetworks}}, \url{http://neo4j.com/docs/stable/rest-api-transactional.html}
}

