% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNodeOutput.R
\name{formatNodeOutput}
\alias{formatNodeOutput}
\title{Format node information}
\usage{
formatNodeOutput(node, returnas)
}
\arguments{
\item{node}{list of nodes containing node information or node attributes to be formatted}
}
\value{
data frame of node information. Return no data frame if error or found nothing.
}
\description{
format and return node information for \code{\link{fetchNode}}.
}
\note{
maximum no. of nodes returned = 30000 nodes
}
\examples{
#querystring="UNWIND [0,1] AS x WITH x MATCH (node) WHERE ID(node) = x RETURN DISTINCT node"
#node = curlRequest.TRANSACTION(cypher=querystring)
#formatNodeOutput(node)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{fetchNode}}
}

