% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPathOutput.R
\name{formatPathOutput}
\alias{formatPathOutput}
\title{Format relationships}
\usage{
formatPathOutput(url)
}
\arguments{
\item{url}{a string specifying relationship url to be formatted}
}
\value{
relationship information with the following components:

\code{source, target} = node grinn id

\code{sourcename, targetname} = node name

\code{sourcelabel, targetlabel} = node type

\code{sourcexref, targetxref} = node cross references

\code{type} = relationship type

\code{datasource} = relationship resource

\code{properties} = relationship properties

Return empty data frame if error or found nothing.
}
\description{
perform curl request to get relationship information for \code{\link{fetchRelationship}}.
}
\note{
curlRequest.TRANSACTION is used in other functions to improve speed.
}
\examples{
#url = "http://localhost:7474/db/data/node/0/relationships/out"
#result = curlRequest.URL(url)
#df = data.frame(t(sapply(result,c)))
#formatPathOutput(df)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{curlRequest}}, \code{\link{fetchRelationship}}
}

