% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphDb.R
\name{getGraphDb}
\alias{getGraphDb}
\title{Get database location}
\usage{
getGraphDb()
}
\value{
url of the database location.
}
\description{
get database location of the current working envornment.
By default the database lacation is http://grinn.genomecenter.ucdavis.edu:7474/db/data/cypher
which is the human database.
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{setGraphDb}}
}

