% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{getDefault Color}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors.}
}
\description{
getDefault Color
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

