% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_factor.R
\name{guess_factor}
\alias{guess_factor}
\title{Guess Factor and return all the column name of p.}
\usage{
XX
}
\arguments{
\item{p}{the data set p.}
}
\value{
option is all the column names of p.
guess_factor_name is the guessed factor name. only support independent factor now.!!!
}
\description{
Guess factor.
XX
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

