% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo.R
\name{histo}
\alias{histo}
\title{test}
\usage{
histo(p, sample_information_selection, e, x = NULL)
}
\arguments{
\item{e}{should be the e, where rows are samples and columns are compounds.}
}
\description{
test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

