% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo_test.R
\name{hypo_test}
\alias{hypo_test}
\title{hypothesis test}
\usage{
hypo_test(e, p, f, independent_factor_name = NULL,
  repeated_factor_name = NULL, confound = NULL, equal_variance_anova = F)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

