% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_expression_data.R
\name{load_expression_data}
\alias{load_expression_data}
\title{Upload expression data.}
\usage{
load_expression_data(file, ...)
}
\arguments{
\item{file}{an object with a child named datapath, which is the direction of the file input}
}
\description{
fdas
dasfs
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

