% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Data_Exploration.R
\name{PCA_Data_Exploration}
\alias{PCA_Data_Exploration}
\title{PCA}
\usage{
PCA_Data_Exploration(e = NULL, p = NULL, f = NULL,
  scree_Title = "Scree Plot", scree_xlab = paste("PC", PC_x),
  scree_ylab = paste("PC", PC_y), PC_x = 1, PC_y = 2, PC_z = 3,
  score_Title = "Score Plot", score_xlab = paste("PC", PC_x),
  score_ylab = paste("PC", PC_y), color_group = NULL, ellipse = T,
  names_on_scatter = NULL, loading_Title, loading_xlab, loading_ylab)
}
\arguments{
\item{e}{should be the e, where rows are samples and columns are compounds.}
}
\description{
PCA.
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

