% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{PCA_plot}
\usage{
PCA_plot(e, f, p, color, sample_information_selection)
}
\arguments{
\item{e}{should be the e, where rows are samples and columns are compounds.}
}
\description{
PCA_plot
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

