% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWordCloud.R
\name{plotWordCloud}
\alias{plotWordCloud}
\title{plot WordCloud}
\usage{
plotWordCloud(wordtable)
}
\arguments{
\item{wordtable}{data frame of word frequency}
}
\description{
plot WordCloud by overriding wordcloud for using in opencpu app.
}
\examples{
# plotWordCloud("wordtable")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{wordcloud}}
}

