% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInfile.R
\name{readInfile}
\alias{readInfile}
\title{Read a file}
\usage{
readInfile(file, header = TRUE, sep ="\t")
}
\description{
read a file overriding read.csv for GUI
}
\examples{
# Set database location to local server
# readInFile("http://localhost:7474/db/data/cypher")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{read.csv}}
}

