% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGraphDb.R
\name{setGraphDb}
\alias{setGraphDb}
\title{Set database location}
\usage{
setGraphDb(url)
}
\arguments{
\item{url}{of the database location}
}
\description{
set database location for the currently working environment.
}
\examples{
# Set database location to local server
setGraphDb("http://localhost:7474/db/data/")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{getGraphDb}}
}

