% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_basic_statistics.R
\name{stat_basic_statistics}
\alias{stat_basic_statistics}
\title{hypothesis test}
\usage{
stat_basic_statistics(e, f, p, independent_factor_name = NULL,
  repeated_factor_name = NULL)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

