% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_BatchMedianNorm.R
\name{stat_BatchMedianNorm}
\alias{stat_BatchMedianNorm}
\title{stat_BatchMedianNorm}
\usage{
stat_BatchMedianNorm(e, f, p, BatchIndicator)
}
\description{
stat_BatchMedianNorm
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

