% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_boxplot.R
\name{stat_boxplot}
\alias{stat_boxplot}
\title{stat_boxplot}
\usage{
stat_boxplot(e, p, f, independent_factor_name = NULL,
  repeated_factor_name = NULL, compound_name_column_index = 1,
  order_of_factor = NULL)
}
\description{
stat_boxplot
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

