% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_Boxplot_generator.R
\name{stat_Boxplot_generator}
\alias{stat_Boxplot_generator}
\title{hypothesis test}
\usage{
stat_Boxplot_generator(e, p, f, e_ori, p_ori, independent_factor_name = NULL,
  repeated_factor_name = NULL, confound = NULL, equal_variance_anova = F)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

