% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_check_hypo_test.R
\name{stat_check_hypo_test}
\alias{stat_check_hypo_test}
\title{stat_check_hypo_test}
\usage{
stat_check_hypo_test(e, p, f, e_ori, p_ori, independent_factor_name = NULL,
  repeated_factor_name = NULL)
}
\description{
stat_check_hypo_test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

