% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_combine_vector_1by1.R
\name{stat_combine_vector_1by1}
\alias{stat_combine_vector_1by1}
\title{combine value one by one for two vectors}
\usage{
stat_combine_vector_1by1(x, y)
}
\description{
combine value one by one for two vectors
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

