% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cross_merge_vectors.R
\name{stat_cross_merge_vectors}
\alias{stat_cross_merge_vectors}
\title{cross merge two vector}
\usage{
stat_cross_merge_vectors(x, y)
}
\description{
cross merge two vector
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

