% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cure_Dunn_format.R
\name{stat_cure_Dunn_format}
\alias{stat_cure_Dunn_format}
\title{format bad sequence of DunnTest}
\usage{
stat_cure_Dunn_format(x, sudo_matrix, temp)
}
\description{
format bad sequence of DunnTest
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

