% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_data_scaling.R
\name{stat_data_scaling}
\alias{stat_data_scaling}
\title{sample normalization}
\usage{
stat_data_scaling(e_after_transformation, f, p, data_scaling = NULL)
}
\description{
sample normalization
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

