% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_data_transformation.R
\name{stat_data_transformation}
\alias{stat_data_transformation}
\title{sample normalization}
\usage{
stat_data_transformation(e_after_sample_normalization, f, p,
  data_transformation = NULL)
}
\description{
sample normalization
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

