% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_delete_result_column.R
\name{stat_delete_result_column}
\alias{stat_delete_result_column}
\title{stat_delete_result_column}
\usage{
stat_delete_result_column(stat_result, selected_feature_by_check)
}
\description{
stat_delete_result_column
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

