% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_Donut_plot.R
\name{stat_Donut_plot}
\alias{stat_Donut_plot}
\title{Donut_plot}
\usage{
stat_Donut_plot(e, f, p, p_column, selected_sample)
}
\arguments{
\item{e}{should be the e, where rows are samples and columns are compounds.}
}
\description{
Donut_plot
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

