% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_get_index_from_PCA.R
\name{stat_get_index_from_PCA}
\alias{stat_get_index_from_PCA}
\title{get Index from PCA}
\usage{
stat_get_index_from_PCA(e, p, f, selected_sample)
}
\description{
get Index from PCA
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

