% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_get_levels.R
\name{stat_get_levels}
\alias{stat_get_levels}
\title{stat_get_levels}
\usage{
stat_get_levels(e, f, p, factor)
}
\description{
stat_get_levels
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

