% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_guess_factor.R
\name{stat_guess_factor}
\alias{stat_guess_factor}
\title{stat_guess_factor}
\usage{
stat_guess_factor(DATA)
}
\description{
stat_guess_factor
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

