% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hypo_test.R
\name{stat_hypo_test}
\alias{stat_hypo_test}
\title{hypothesis test}
\usage{
stat_hypo_test(e, p, f, e_ori, p_ori, e_before, p_before,
  independent_factor_name = NULL, repeated_factor_name = NULL,
  need_power = T, desired_power = NULL, ttestmethod = NULL,
  ttestcorrection = NULL, nonparattestmethod = NULL,
  nonparattestcorrection = NULL, ANOVAmethod = NULL, ANOVAposthoc = NULL,
  nonparaANOVAmethod = NULL, nonparaANOVAposthoc = NULL,
  twowayANOVAmethod = NULL, maineffect1ANOVAmethod = NULL,
  maineffect1ANOVAposthoc = NULL, simplemaineffect1ANOVAmethod = NULL,
  simplemaineffect1ANOVAposthoc = NULL, maineffect1ttestmethod = NULL,
  maineffect1ttestcorrection = NULL, simplemaineffect1ttestmethod = NULL,
  simplemaineffect1ttestcorrection = NULL, maineffect2ANOVAmethod = NULL,
  maineffect2ANOVAposthoc = NULL, simplemaineffect2ANOVAmethod = NULL,
  simplemaineffect2ANOVAposthoc = NULL, maineffect2ttestmethod = NULL,
  maineffect2ttestcorrection = NULL, simplemaineffect2ttestmethod = NULL,
  simplemaineffect2ttestcorrection = NULL, nonparatwowayANOVAmethod = NULL,
  nonparamaineffect1ANOVAmethod = NULL,
  nonparamaineffect1ANOVAposthoc = NULL,
  nonparasimplemaineffect1ANOVAmethod = NULL,
  nonparasimplemaineffect1ANOVAposthoc = NULL,
  nonparamaineffect1ttestmethod = NULL,
  nonparamaineffect1ttestcorrection = NULL,
  nonparasimplemaineffect1ttestmethod = NULL,
  nonparasimplemaineffect1ttestcorrection = NULL,
  nonparamaineffect2ANOVAmethod = NULL,
  nonparamaineffect2ANOVAposthoc = NULL,
  nonparasimplemaineffect2ANOVAmethod = NULL,
  nonparasimplemaineffect2ANOVAposthoc = NULL,
  nonparamaineffect2ttestmethod = NULL,
  nonparamaineffect2ttestcorrection = NULL,
  nonparasimplemaineffect2ttestmethod = NULL,
  nonparasimplemaineffect2ttestcorrection = NULL, pairedttestmethod = NULL,
  pairedttestcorrection = NULL, nonparapairedttestmethod = NULL,
  nonparapairedttestcorrection = NULL, pairedANOVAmethod = NULL,
  pairedANOVAadjust = NULL, pairedANOVAposthoc = NULL,
  nonparapairedANOVAmethod = NULL, nonparapairedANOVAposthoc = NULL,
  twowaypairedANOVAmethod = NULL, twowaypairedANOVAadjust = NULL,
  maineffect1pairedANOVAmethod = NULL, maineffect1pairedANOVAadjust = NULL,
  maineffect1pairedANOVAposthoc = NULL,
  simplemaineffect1pairedANOVAmethod = NULL,
  simplemaineffect1pairedANOVAadjust = NULL,
  simplemaineffect1pairedANOVAposthoc = NULL,
  maineffect1pairedttestmethod = NULL,
  maineffect1pairedttestcorrection = NULL,
  simplemaineffect1pairedttestmethod = NULL,
  simplemaineffect1pairedttestcorrection = NULL,
  maineffect2pairedANOVAmethod = NULL, maineffect2pairedANOVAadjust = NULL,
  maineffect2pairedANOVAposthoc = NULL,
  simplemaineffect2pairedANOVAmethod = NULL,
  simplemaineffect2pairedANOVAadjust = NULL,
  simplemaineffect2pairedANOVAposthoc = NULL,
  maineffect2pairedttestmethod = NULL,
  maineffect2pairedttestcorrection = NULL,
  simplemaineffect2pairedttestmethod = NULL,
  simplemaineffect2pairedttestcorrection = NULL,
  nonparatwowaypairedANOVAmethod = NULL,
  nonparamaineffect1pairedANOVAmethod = NULL,
  nonparamaineffect1pairedANOVAposthoc = NULL,
  nonparasimplemaineffect1pairedANOVAmethod = NULL,
  nonparasimplemaineffect1pairedANOVAposthoc = NULL,
  nonparamaineffect1pairedttestmethod = NULL,
  nonparamaineffect1pairedttestcorrection = NULL,
  nonparasimplemaineffect1pairedttestmethod = NULL,
  nonparasimplemaineffect1pairedttestcorrection = NULL,
  nonparamaineffect2pairedANOVAmethod = NULL,
  nonparamaineffect2pairedANOVAposthoc = NULL,
  nonparasimplemaineffect2pairedANOVAmethod = NULL,
  nonparasimplemaineffect2pairedANOVAposthoc = NULL,
  nonparamaineffect2pairedttestmethod = NULL,
  nonparamaineffect2pairedttestcorrection = NULL,
  nonparasimplemaineffect2pairedttestmethod = NULL,
  nonparasimplemaineffect2pairedttestcorrection = NULL, mixedANOVA = NULL,
  mixedANOVAadjust = NULL, nonparamixedANOVA = NULL)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

