% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_LoessNorm.R
\name{stat_LoessNorm}
\alias{stat_LoessNorm}
\title{stat_LoessNorm}
\usage{
stat_LoessNorm(data = e, f, p, loess.para = 0.75,
  auto.batch.detection = FALSE, robust = TRUE)
}
\description{
stat_LoessNorm
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

