% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_median_correction.R
\name{stat_median_correction}
\alias{stat_median_correction}
\title{stat_median_correction}
\usage{
stat_median_correction(before, batch, QC)
}
\description{
stat_median_correction
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

