% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mixed_ANOVA_power.R
\name{stat_mixed_ANOVA_power}
\alias{stat_mixed_ANOVA_power}
\title{stat_mixed_ANOVA_power}
\usage{
stat_mixed_ANOVA_power(e, p, dta, sig.level = 0.05, desired_power = 0.8,
  factor_name, epsilon = 1, cl)
}
\arguments{
\item{result_stat:}{it must be the basic statistics result.}
}
\description{
stat_mixed_ANOVA_power
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

