% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_norm.R
\name{stat_norm}
\alias{stat_norm}
\title{normalization}
\usage{
stat_norm(e, f, p, sample_index, mTICdid, mTICunchanged, Loessdid,
  Loessunchanged, medFCdid, BatchMediandid, mTIC, Loess, medFC, BatchMedian,
  sample_normalization = NULL, data_transformation = NULL,
  data_scaling = NULL, sample_specific_weight = NULL,
  sample_specific_multiplyordevide = "Multiply", KnownorUnknown = NULL,
  QCIndicator = NULL, BatchIndicator = NULL, TimeIndicator = NULL,
  Batchunchanged, log_para, independent_factor_name_log, power_para,
  independent_factor_name_power)
}
\description{
normalization
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

