% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_one_way_ANOVA.R
\name{stat_one_way_ANOVA}
\alias{stat_one_way_ANOVA}
\title{normalization}
\usage{
stat_one_way_ANOVA(data, data2, i, sudo_matrix, factor_name, cl, ANOVAmethod,
  ANOVAposthoc, nonparaANOVAmethod, nonparaANOVAposthoc)
}
\description{
normalization
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

