% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_one_way_repeated_ANOVA.R
\name{stat_one_way_repeated_ANOVA}
\alias{stat_one_way_repeated_ANOVA}
\title{stat_one_way_repeated_ANOVA}
\usage{
stat_one_way_repeated_ANOVA(data, data2, i, sudo_matrix, factor_name, cl,
  onewaySpher_Corr)
}
\description{
stat_one_way_repeated_ANOVA
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

