% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_paired_t_test.R
\name{stat_paired_t_test}
\alias{stat_paired_t_test}
\title{stat_paired_t_test}
\usage{
stat_paired_t_test(data, data2, i, cl)
}
\description{
stat_paired_t_test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

