% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pareto_scale.R
\name{stat_pareto_scale}
\alias{stat_pareto_scale}
\title{stat_pareto_scale}
\usage{
stat_pareto_scale(e, f, p)
}
\description{
stat_pareto_scale
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

