% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_PCA_plot.R
\name{stat_PCA_plot}
\alias{stat_PCA_plot}
\title{stat_PCA_plot}
\usage{
stat_PCA_plot(e, f, p, color, sample_information_selection = NULL,
  Ellipse = TRUE, dot_size = 18, IDtrend = T)
}
\arguments{
\item{e}{should be the e, where rows are samples and columns are compounds.}
}
\description{
stat_PCA_plot
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

