% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_range_scale.R
\name{stat_range_scale}
\alias{stat_range_scale}
\title{stat_range_scale}
\usage{
stat_range_scale(e, f, p)
}
\description{
stat_range_scale
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

