% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_rm_sample.R
\name{stat_rm_sample}
\alias{stat_rm_sample}
\title{stat_rm_sample}
\usage{
stat_rm_sample(e, p, f, sample_index)
}
\description{
stat_rm_sample
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

