% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sample_norm.R
\name{stat_sample_norm}
\alias{stat_sample_norm}
\title{sample normalization}
\usage{
stat_sample_norm(e, f, p, sample_normalization = NULL)
}
\description{
sample normalization
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

