% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary_data.R
\name{stat_summary_data}
\alias{stat_summary_data}
\title{Summary Data}
\usage{
load_aggregated_data(file, type, ...)
}
\arguments{
\item{file}{the file in read.csv or read.xlsx2.}

\item{type}{a string of file name ended either with .xlsx or .csv.}

\item{...}{Additional arguments for xlsx::read.xlsx2 or read.csv.}
}
\value{
a list of three data frames: "expression"(sample in row), "feature"(compoud in row) and "phenotype"(sample in row).
}
\description{
Summary Data
}
\details{

}
\examples{
load_aggregated_data(input$inputID,startRow=2)
}
\author{
Sili Fan \email{fansili2013@gmail.com}
}
\seealso{
\code{\link{load_expression_data}}, \code{\link{load_expression_data}}, \code{\link{load_expression_data}}
}

