% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_t_test.R
\name{stat_t_test}
\alias{stat_t_test}
\title{normalization}
\usage{
stat_t_test(data, data2, i, cl, ttestmethod, ttestFDRmethod,
  nonpara_ttestmethod, nonpara_ttestFDRmethod)
}
\description{
normalization
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

