% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_t_test_power.R
\name{stat_t_test_power}
\alias{stat_t_test_power}
\title{stat_t_test_power}
\usage{
stat_t_test_power(dta, i, result_stat, sig.level = 0.05,
  desired_power = 0.8, independent_factor_name, cl)
}
\arguments{
\item{result_stat:}{it must be the basic statistics result.}
}
\description{
stat_t_test_power
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

