% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test_boxplot.R
\name{stat_test_boxplot}
\alias{stat_test_boxplot}
\title{hypothesis test}
\usage{
stat_test_boxplot(n)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

