% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test_outlier.R
\name{stat_test_outlier}
\alias{stat_test_outlier}
\title{test outleir}
\usage{
stat_test_outlier(e, f, p, class)
}
\description{
test outlier
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

