% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_two_way_ANOVA_power.R
\name{stat_two_way_ANOVA_power}
\alias{stat_two_way_ANOVA_power}
\title{stat_two_way_ANOVA_power}
\usage{
stat_two_way_ANOVA_power(e, dta, sig.level = 0.05, desired_power = 0.8,
  independent_factor_name, cl)
}
\arguments{
\item{result_stat:}{it must be the basic statistics result.}
}
\description{
stat_two_way_ANOVA_power
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

