% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_aggregated_data.R
\name{summary_aggregated_data}
\alias{summary_aggregated_data}
\title{summarize aggregated data.}
\usage{
summary_aggregated_data(aggregated_data, factor_name, repeated.factor_name)
}
\arguments{
\item{factor_name}{The factor names that are related to the study design.
factor_name must be one of the column name of the "phenotype" dataframe. Although the default is NULL, it shouldn't be NULL for mETABOX
statistical analysis.}

\item{aggregated_data}{an list with three dataframes named as "expression", "feature" and "phenotype". Or simply a object returned from
load_aggregated_data().}

\item{repeated.factor_name}{The repeated factor name indicates which factor is repeated measure. Must be one of the factor_name.
If NULL means that it is independed study design.}
}
\value{
a list. If the input aggregated_data is not standard, then it returns a warning message telling user to upload needed file.
dataset(list), factor_name(vector of string), repeated.factor_name(vector of string), confound(vector which is a variable),
batch(vector which is a variable of factor)
}
\description{
summarize aggregated data.
Summarize the information that is important for mETABOX statistical analysis, e.g. factor_name, repeated.factor_name, etc.
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

