% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test(e, f, p, color, sample_information_selection)
}
\arguments{
\item{e}{should be the e, where rows are samples and columns are compounds.}
}
\description{
test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

