% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testing}
\alias{testing}
\title{a}
\usage{
testing()
}
\value{
a
}
\description{
a
Summarize the information that is important for mETABOX statistical analysis, e.g. factor_name, repeated.factor_name, etc.
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

