% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callBionet.R
\name{callBionet}
\alias{callBionet}
\title{Compute a maximum scoring subnetwork}
\usage{
callBionet(edgelist, nodelist, pval, fdr)
}
\arguments{
\item{edgelist}{a data frame of edges contains at least a source column (1st column) and a target column (2nd column).}

\item{nodelist}{a data frame of nodes contains node attributes e.g. node id, node name, node xref. Default is NULL.}

\item{pval}{a numeric vector of p-values with name attributes identical to the names of network nodes.}

\item{fdr}{a numeric value specifying false discovery rate. Default is 0.05.}
}
\value{
list of data frame of nodes and data frame of edges with the following components:

nodes:

\code{nodeAttributes} = node attributes provided

\code{score} = node score: positive values = signal content and negative values = background noise

edges:

\code{source, target}

\code{edgeAttributes} = edge attributes provided

Return list of empty data frame if error or found nothing.
}
\description{
compute a maximum scoring subnetwork from p-values of the nodes of a network.
The function wraps around the main steps of \pkg{\link{BioNet}} including \code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}
to compute a subnetwork.
}
\examples{
#simnw <- computeSimilarity(c(1110,10413,196,51,311,43,764,790)) #compute similarity network for given pubchem compounds
#pval <- data.frame(pubchem=c(1110,10413,196,51,311,43,764,790),stat=runif(8, 0, 0.06)) #statistical values of pubchem compounds
#result <- computeSubnetwork(sim$edges,sim$nodes,pval=pval,internalid = F)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Beisser D., Klau GW., Dandekar T., Müller T. and Dittrich MT. (2010) BioNet: an R-Package for the functional analysis of biological networks. Bioinformatics, 26(8):1129-30

Dittrich MT., Klau GW., Rosenwald A., Dandekar T., Müller T. (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. Bioinformatics, 24(13):i223-31
}
\seealso{
\code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}
}

