% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWordCloud.R
\name{callWordCloud}
\alias{callWordCloud}
\title{Compute wordcloud for a given annotation}
\usage{
callWordCloud(edgelist, nodelist)
}
\arguments{
\item{edgelist}{a data frame of edges}

\item{nodelist}{a data frame of nodes containin node information e.g. node id, node gid, node name, node xref.}
}
\value{
data frame of word frequency. Return empty data frame if error or found nothing.
}
\description{
compute wordcloud for a given annotation retrieved from e.g. \code{\link{fetchNetwork}}.
}
\examples{
# pc = read.csv("~/Desktop/testGrinn/pubchemls.txt")
# nw = fetchNetworkByGID(to=pc$GrinnID,fromtype = "pathway",totype = "compound",reltype = "annotation")
# wd = callWordCloud(nw$edges,nw$nodes)
# wordcloud::wordcloud(words = gsub(" - Mus musculus \\\\(mouse\\\\)","",wd$nodename), freq = wd$freq, scale=c(2,.1),min.freq = 1,max.words=50, random.order=FALSE, rot.per=0.5, colors=brewer.pal(8, "Dark2"))
# barplot(wd$freq[1:10], las = 2, names.arg = gsub(" - Mus musculus \\\\(mouse\\\\)","",wd$nodename[1:10]), col ="lightblue", main ="Most frequent words", ylab = "Word frequencies")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{fetchNetwork}}, \code{\link{fetchHetNetwork}}, \pkg{\link{tm}}, \pkg{\link{wordcloud}}
}

