% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineNetworks.R
\name{combineNetworks}
\alias{combineNetworks}
\title{Combine networks}
\usage{
combineNetworks(networks, returnas)
}
\arguments{
\item{networks}{list of networks. Each network contains data frame of nodes and data frame of edges.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of nodes and edges. Return empty list if error or found nothing.
}
\description{
The function can combine networks computed from e.g. \code{\link{fetchNetwork}}, \code{\link{fetchHetNetwork}}, \code{\link{computeCorrelation}},
\code{\link{computeParCorrelation}}, and \code{\link{computeSimilarity}}.
}
\examples{
#simnw <- computeSimilarity(c(1110,10413,196,51,311,43,764,790)) #compute similarity network for given pubchem compounds
#dt <- data.frame(id=seq(32), mtcars, row.names = NULL) #data frame of x
#cornw <- computeCorrelation(x = dt, xtype = "compound", coef=0.9, pval = 1e-10) #compute correlation network
#result <- combineNetworks(list(simnw, cornw))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{fetchNetwork}}, \code{\link{fetchHetNetwork}}, \code{\link{computeCorrelation}}, \code{\link{computeParCorrelation}}, \code{\link{computeSimilarity}}
}

