% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelation.R
\name{computeCorrelation}
\alias{computeCorrelation}
\title{Compute a weighted correlation network}
\usage{
computeCorrelation(x, y, xtype, ytype, internalid, coef, pval, method, returnas)
}
\arguments{
\item{x}{a data frame of quantified omic data e.g. gene expression data, metabolite intensities.
Columns are samples and rows are molecular entities e.g. genes, proteins or compounds.}

\item{y}{a data frame of quantified omic data e.g. gene expression data, metabolite intensities.
Columns are samples and rows are molecular entities e.g. genes, proteins or compounds.}

\item{xtype}{a string specifying the type of nodes (default = NULL). It can be one of compound, protein, gene, rna, dna.}

\item{ytype}{a string specifying the type of nodes (default = NULL). It can be one of compound, protein, gene, rna, dna.}

\item{internalid}{boolean value, whether name attributes of pval are neo4j ids, see \code{\link{convertId}} for how to convert to neo4j ids.}

\item{coef}{a numeric value specifying the minimum absolute correlation coefficient to be included in the output (from 0 to 1, default is 0.7).}

\item{pval}{a numeric value specifying the maximum p-value to be included in the output (default is 0.05).}

\item{method}{a string specifying method to compute correlation. It can be one of pearson, kendall, spearman (default) see \code{\link{cor}}.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of nodes with the following components:

\code{id} = node id

\code{gid} = node gid

\code{nodename} = node name

edgelist with the following components:

\code{source, target} = node id

\code{coef} = correlation coefficient

\code{pval} = p-value

\code{direction} = correlation direction

\code{type} = relationship type

Return empty list or data frame if error or found nothing.
}
\description{
compute the correlation network for one or two omic data sets.
The function wraps around \code{\link{corAndPvalue}} function of \pkg{\link{WGCNA}}.
Correlation coefficients, p-values and correlation directions are calculated.
The correlation coefficients are continuous values between -1 (negative correlation) and 1 (positive correlation), with numbers close to 1 or -1, meaning very closely correlated.
}
\details{
x and y are data frame in which columns are samples and rows are molecular entities e.g. genes, proteins or compounds.
If y is given, then the correlations between the molecular entities of x and of y are computed. Otherwise the correlations between molecular entities of x are computed.
If grinn functions will be used further, we recommend using the grinn ids for entities, see \code{\link{convertId}}.
}
\examples{
# Compute a correlation network of metabolites
#dummy <- rbind(nodetype=rep("metabolite"),t(mtcars))
#colnames(dummy) <- c('G1.1','G27967','G371','G4.1',paste0('G',sample(400:22000, 28)))
#result <- fetchCorrNetwork(datNormX=dummy, datNormY=NULL, corrCoef=0.7, pval=1e-12, method="spearman", returnAs="tab")
#library(igraph)
#plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
# Compute a correlation network of metabolites and proteins
#dummyX <- rbind(nodetype=rep("metabolite"),t(mtcars)[,1:16])
#colnames(dummyX) <- c('G1.1','G27967','G371','G4.1',paste0('G',sample(400:22000, 12)))
#dummyY <- rbind(nodetype=rep("protein"),t(mtcars)[,17:32])
#colnames(dummyY) <- c('P28845','P08235','Q08AG9','P80365',paste0('P',sample(10000:80000, 12)))
dummyX <- mtcars[1:16,]
dummyY <- mtcars[17:32,]
#result <- computeCorrelation(x=dummyX, y=dummyY, coef=0.7, pval=1e-4, method="spearman", returnas="dataframe")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Langfelder P. and Horvath S. (2008) WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics, 9:559

Dudoit S., Yang YH., Callow MJ. and Speed TP. (2002) Statistical methods for identifying differentially expressed genes in replicated cDNA microarray experiments, STATISTICA SINICA, 12:111

Langfelder P. and Horvath S. Tutorials for the WGCNA package \url{http://labs.genetics.ucla.edu/horvath/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/index.html}
}
\seealso{
\code{\link{cor}}, \code{\link{corAndPvalue}}
}

