% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnrichment.R
\name{computeEnrichment}
\alias{computeEnrichment}
\title{Perform enrichment analysis}
\usage{
computeEnrichment(edgelist, pval, fc, method, size, returnas)
}
\arguments{
\item{edgelist}{a two-column data frame of annotation pairs. The 1st column contains annotated entities and the 2nd column contains annotation terms. See \code{\link{loadGSC}} for details.}

\item{pval}{a numeric vector of statistical values e.g. p-values. The name attributes must be identical to the names of entities. See \code{\link{runGSA}} for details.}

\item{fc}{a numeric vector of fold changes or sign information (positive or negative) with name attributes identical to the names of entities. See \code{\link{runGSA}} for details. Default is NULL.}

\item{method}{a string specifying the enrichment analysis method. It can be one of reporter (default), fisher, median, mean, stouffer. See \code{\link{runGSA}}}

\item{size}{a numeric vector specifying the minimum and maximum number of members in each annotation term to be used in the analysis. Default is c(3,500).}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
enrichment analysis result with the following components:

\code{id} = annotation id

\code{no_of_entities} = number of members in each annotation term

\code{p} = raw p-values

\code{p_adj} = adjusted p-values

\code{member} = list of entity members of the annotation term

Return empty list or data frame if error or found nothing.
}
\description{
perform enrichment analysis from p-values of entities. The function wraps around the main functions of \pkg{\link{piano}}.
}
\examples{
#simnw <- computeSimilarity(c(1110,10413,196,51,311,43,764,790)) #compute similarity network for given pubchem compounds
#pval <- data.frame(pubchem=c(1110,10413,196,51,311,43,764,790), stat=runif(8, 0, 0.06)) #statistical values of pubchem compounds
#result <- computeNwEnrichment(simnw$edges, simnw$nodes, annotation="mesh", pval, internalid = FALSE)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Fisher R. (1932) Statistical methods for research workers. Oliver and Boyd, Edinburgh.

Stouffer S., Suchman E., Devinney L., Star S., and Williams R. (1949) The American soldier: adjustment during army life. Princeton University Press, Oxford, England.

Patil K. and Nielsen J. (2005) Uncovering transcriptional regulation of metabolism by using metabolic network topology. Proceedings of the National Academy of Sciences of the United States of America 102(8), 2685.

Oliveira A., Patil K., and Nielsen J. (2008) Architecture of transcriptional regulatory circuits is knitted over the topology of bio-molecular interaction networks. BMC Systems Biology 2(1), 17.

Väremo L., Nielsen J., and Nookaew I. (2013) Enriching the gene set analysis of genome-wide data by incorporating directionality of gene expression and combining statistical hypotheses and methods. Nucleic Acids Research, 41(8), pp. 4378-4391.
}
\seealso{
\code{\link{loadGSC}}, \code{\link{runGSA}}, \code{\link{GSAsummaryTable}}
}

