% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNodeEnrichment.R
\name{computeNodeEnrichment}
\alias{computeNodeEnrichment}
\title{Perform enrichment analysis on the input entities}
\usage{
computeNodeEnrichment(nodedata, nodetype, annotation, internalid, method, size)
}
\arguments{
\item{nodedata}{a two-column data frame of entities and the statistical values. The 1st column is entities and the second is statistical values e.g. p-values.}

\item{nodetype}{a string specifying a node type. It can be one of compound (default), protein, gene, pathway, rna, dna.}

\item{annotation}{a string specifying the annotation type e.g. pathway (default) and mesh. Pathway annotation requires the database.
Mesh annotation doesn't require the database but it is available for PubChem compounds only.}

\item{internalid}{a logical value indicating whether the input entities are the neo4j ids, if TRUE (default). If not, the entities are expected to be the grinn id or PubChem CID.
See details and see \code{\link{convertId}} for how to convert ids. It has no effect on Mesh annotation.}

\item{method}{a string specifying the enrichment analysis method. It can be one of reporter (default), fisher, median, mean, stouffer. See \code{\link{runGSA}}}

\item{size}{a numeric vector specifying the minimum and maximum number of members in each annotation term to be used in the analysis. Default is c(3,500).}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}

\item{method}{a string specifying the enrichment analysis method. It can be one of reporter (default), fisher, median, mean, stouffer. See \code{\link{runGSA}}}

\item{size}{a numeric vector specifying the minimum and maximum number of members in each annotation term to be used in the analysis. Default is c(3,500).}
}
\value{
list of data frame of nodes, edges and enrichment. The nodes data frame contains input attributes. The edges data frame contains annotation pairs.
The data frame of enrichment contains the following components:

list of data frame of nodes, edges, enrichment and pairs. The pairs data frame contains annotation pairs. The data frame of enrichment contains the following components:

\code{rank} = rank sort by p

\code{id} = annotation id or annotation neo4j id

\code{gid} = annotation id or annotation grinn id

\code{nodename} = annotation name

\code{nodelabel} = annotation type

\code{nodexref} = cross references

\code{p} = raw p-values

\code{p_adj} = adjusted p-values

\code{no_of_entities} = number of input entities in each annotation term

\code{annotation_size} = total number of entities in each annotation term from the database

\code{member} = list of entity members of the annotation term

Return list of empty data frame if error or found nothing.
}
\description{
perform enrichment analysis from p-values of the input entities. The function wraps around the main functions of \pkg{\link{piano}}.
}
\details{
The database uses two id systems. The neo4j id is a numeric, internal id automatically generated by the database system.
The grinn id (gid) is an id system of Grinn database that uses main ids of standard resources
i.e. ENSEMBL for genes (e.g.ENSG00000139618), UniProt for proteins (e.g.P0C9J6), PubChem CID for compounds (e.g.5793), KEGG for pathways (e.g.hsa00010).

The database uses two id systems. The neo4j id is a numeric, internal id automatically generated by the database system.
The grinn id (gid) is an id system of Grinn database that uses main ids of standard resources
i.e. ENSEMBL for genes (e.g.ENSG00000139618), UniProt for proteins (e.g.P0C9J6), PubChem CID for compounds (e.g.5793), KEGG for pathways (e.g.hsa00010).
}
\examples{
#dt <- data.frame(pubchem=c(1110,10413,196,51,311,43,764,790), stat=runif(8, 0, 0.06)) #statistical values of pubchem compounds
#result <- computeNodeEnrichment(nodedata=dt, nodetype="compound", annotation="mesh", internalid=FALSE)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Fisher R. (1932) Statistical methods for research workers. Oliver and Boyd, Edinburgh.

Stouffer S., Suchman E., Devinney L., Star S., and Williams R. (1949) The American soldier: adjustment during army life. Princeton University Press, Oxford, England.

Patil K. and Nielsen J. (2005) Uncovering transcriptional regulation of metabolism by using metabolic network topology. Proceedings of the National Academy of Sciences of the United States of America 102(8), 2685.

Oliveira A., Patil K., and Nielsen J. (2008) Architecture of transcriptional regulatory circuits is knitted over the topology of bio-molecular interaction networks. BMC Systems Biology 2(1), 17.

Väremo L., Nielsen J., and Nookaew I. (2013) Enriching the gene set analysis of genome-wide data by incorporating directionality of gene expression and combining statistical hypotheses and methods. Nucleic Acids Research, 41(8), pp. 4378-4391.
}
\seealso{
\code{\link{loadGSC}}, \code{\link{runGSA}}, \code{\link{GSAsummaryTable}}
}

