% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeParCorrelation.R
\name{computeParCorrelation}
\alias{computeParCorrelation}
\title{Compute partial correlation network}
\usage{
computeParCorrelation(x, xtype, internalid, coef, pval, alpha, epsilon, matrix.completion, returnas)
}
\arguments{
\item{x}{a data frame of raw or quantified data e.g. gene expression data, metabolite intensities.
Columns are samples and rows are entities e.g. genes, proteins or compounds.}

\item{xtype}{a string specifying a node type (default = NULL). If provided and the database is installed, node attributes will be automatically retrieved from the database.
For database query, the value can be one of compound, protein, gene, rna, dna.}

\item{internalid}{a logical value indicating whether the network nodes are neo4j ids, if TRUE (default). See \code{\link{convertId}} for how to convert ids.
It has no effect if \code{xtype} = NULL or there is no database installed.}

\item{coef}{a numeric value specifying the minimum absolute partial correlation coefficient to be included in the output (from 0 to 1, default is 0.7).}

\item{pval}{a numeric value specifying the maximum p-value to be included in the output (default is 0.05).}

\item{alpha}{a numeric value specifying significance level of each test used in \code{\link{qpAvgNrr}}.}

\item{epsilon}{a numeric value specifying the maximum cutoff value of the non-rejection rate met by the edges that are included in the qp-graph, see \code{\link{qpGraph}}.}

\item{matrix.completion}{a string specifying algorithm to employ in the matrix completion operations used in \code{\link{qpPAC}}}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of network information with the following components:

nodes:

\code{id} = node id or node neo4j id

\code{gid} = node id or node grinn id

\code{nodename} =  node id or node name

\code{nodelabel} = node type if provided

edges:

\code{source, target} = node id or node neo4j id

\code{coef} = partial correlation coefficient

\code{pval} = p-value

\code{direction} = correlation direction

\code{type} = relationship type

Return empty list if error or found nothing.
}
\description{
compute the partial correlation network of entities from raw or quantified data, see details.
}
\details{
The function wraps around the functions of \pkg{\link{qpgraph}}.
Partial correlation coefficients, p-values and correlation directions are calculated.
The partial correlation coefficients are continuous values between -1 (negative correlation) and 1 (positive correlation), with numbers close to 1 or -1, meaning very closely correlated.
}
\note{
If the database is installed, node attributes will be automatically retrieved from the database. Otherwise node attributes will be the original input.
}
\examples{
#dt <- data.frame(id=row.names(mtcars), mtcars, row.names = NULL) #data frame of x
#result <- computeParCorrelation(x = dt)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Castelo R. and Roverato A. (2006) A robust procedure for Gaussian graphical model search from microarray data with p larger than n. J. Mach. Learn. Res., 7:2621-2650.

Castelo R. and Roverato A. (2009) Reverse engineering molecular regulatory networks from microarray data with qp-graphs. J Comput Biol, 16(2), pp. 213-27.
}
\seealso{
\pkg{\link{qpgraph}}, \code{\link{qpAvgNrr}}, \code{\link{qpGraph}}, \code{\link{qpPAC}}
}

