% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{computeSimilarity}
\alias{computeSimilarity}
\title{Compute Tanimoto similarity network}
\usage{
computeSimilarity(txtinput, coef, returnas)
}
\arguments{
\item{txtinput}{a character vector of PubChem CIDs given for structure similarity comparisons.}

\item{coef}{a numeric value specifying the minimum Tanimoto similarity correlation coefficient to be included in the output (from 0 to 1, default is 0.7).}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of nodes with the following components:

\code{id} = node internal neo4j id

\code{gid} = node grinn id

\code{nodename} = node name

\code{nodelabel} = node type

edgelist with the following components:

\code{source, target} = node

\code{coef} = Tanimoto similarity coefficient

\code{type} = relationship type

Return empty list or data frame if error or found nothing.
}
\description{
compute Tanimoto similarity for compounds. 
The function wraps around the functions of \pkg{\link{metabomapr}} to compute Tanimoto distances of all given compounds using PubChem fingerprints.
}
\examples{
#datNorm = read.csv("~/Documents/grinn_sample/lung_miyamoto/metAdj.txt",sep="\\t",header=TRUE,row.names=1)
#datNorm = datNorm[-1,] #exclude nodetype
#dt = sapply(datNorm, function(x) as.numeric(as.character(x)))
#nw = fetchPtCorrNetwork(datNorm=dt)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Willett P., Barnard JM. and Downs GM. (1998) Chemical similarity searching. J. Chem. Inf. Comput. Sci., 38, 983–996.

Grapov D., Wanichthanarak K. and Fiehn O. (2015) MetaMapR: pathway independent metabolomic network analysis incorporating unknowns. Bioinformatics. 31(16):2757-60.

Cao Y., Charisi A., Cheng L., Jiang T. and Girke T. (2008) ChemmineR: a compound mining framework for R. Bioinformatics, 24(15), pp. 1733–1734.

\url{ftp://ftp.ncbi.nih.gov/pubchem/specifications/pubchem_fingerprints.txt}
}
\seealso{
\pkg{\link{metabomapr}}, \code{\link{CID_tanimoto}}, \code{\link{fpSim}}, \code{\link{sdfid}}
}

