% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSubnetwork.R
\name{computeSubnetwork}
\alias{computeSubnetwork}
\title{Compute a subnetwork}
\usage{
computeSubnetwork(edgelist, nodelist, pval, fc, fdr, internalid, method, returnas)
}
\arguments{
\item{edgelist}{a data frame of edges contains at least source and target pairs}

\item{nodelist}{a data frame of nodes contains node information e.g. node name, node xref. Default is NULL.}

\item{pval}{a numeric vector of p-values with name attributes corresponding to network nodes}

\item{fc}{a numeric vector of fold changes with name attributes corresponding to network nodes}

\item{fdr}{a numeric value specifying false discovery rate. Default is 0.05.}

\item{internalid}{boolean value, whether name attributes of pval are neo4j ids, see \code{\link{convertId}} for how to convert to neo4j ids.}

\item{method}{a string specifying method used to compute the subnetwork. It can be one of bionet, frank. Default is bionet. See \code{details}}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
list of nodes and edges. The list contains the data frame of nodes and the data frame of edges. colnames of nodes = name (igraph default), ..., score
Return empty list of data frame if error or found nothing.
}
\description{
compute a subnetwork or module by integrating omic data and the biological network.
The function wraps around the main functions of \pkg{\link{BioNet}} including \code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}.
}
\details{
The method \code{bionet} encapsulates the main functions from \pkg{\link{BioNet}} to compute a functional module or significanly changed subnetworks.
Those functions include: \code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}.
}
\examples{
#kw <- c('G15729','G17561','G16015','G18145','G16708')
#grinnNW <- fetchGrinnNetwork(txtInput=kw, from="metabolite", to="protein")
#library(grinn)
#data(dummyStat)
#result <- fetchSubnetwork(statInput = dummyStat, nwInput = grinnNW)
#library(igraph)
#plot(graph.data.frame(result$edges[,1:2], directed=FALSE))
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
Beisser D., Klau GW., Dandekar T., Müller T. and Dittrich MT. (2010) BioNet: an R-Package for the functional analysis of biological networks. Bioinformatics, 26(8):1129-30

Dittrich MT., Klau GW., Rosenwald A., Dandekar T., Müller T. (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. Bioinformatics, 24(13):i223-31
}
\seealso{
\code{\link{fitBumModel}}, \code{\link{scoreNodes}}, \code{\link{runFastHeinz}}
}

