% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlRequest.R
\name{curlRequest}
\alias{curlRequest}
\title{Execute query}
\usage{
curlRequest(param)
}
\arguments{
\item{cypher}{a string of Cypher (default) or url a string of url}
}
\value{
list or data frame. Return no data if found nothing.
}
\description{
execute query using Cypher or url as an input. The function calls curlPerform from \pkg{RCurl} for sending HTTP request to the graph database.
}
\examples{
# Query metabolites by database id
#querystring <- "UNWIND ['MT-TF','MT-ND3'] AS x WITH x MATCH (node:Gene) WHERE node.name = x RETURN DISTINCT node"
#result <- curlRequest(querystring)
#url <- "http://localhost:7474/db/data/node/306/relationships/out"
#result <- curlRequest.URL(url)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\references{
The RCurl package \url{http://www.inside-r.org/packages/cran/RCurl}
}
\seealso{
\code{\link{curlPerform}}, \url{http://neo4j.com/docs/milestone/introduction.html}
}

