% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchNode.R
\name{fetchNode}
\alias{fetchNode}
\title{Query node(s) from the graph database}
\usage{
fetchNode(txtinput, nodetype, searchby, exactmatch, returnas)
}
\arguments{
\item{txtinput}{a character vector of values used for the query e.g. txtinput = c('name1', 'name2'). 
The value can be a kind of node property keys, see \code{searchby}. Default is neo4jid.}

\item{nodetype}{a string specifying the type of a query node. It can be one of compound, protein, gene, pathway, rna, dna, biofeature.}

\item{searchby}{a string specifying a node property key used for the query. It can be one of neo4jid, grinnid, name, synonyms, description, properties, xref, datasource. Default is grinnid.}

\item{exactmatch}{logical. If TRUE, it will query for exact results i.e. a case-sensitive and exact keyword or phrase. Default is TRUE.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
node information. Return empty list or data frame if error or found nothing.
}
\description{
query node(s) from the graph database. The results contain node information including neo4jid, grinnid, nodename, 
nodelabel, datasource, description, organism, synonyms and xref, if applicable.
}
\examples{
# Query genes by name
#kw <- list('MT-TF','MT-TQ','MT-TI','MT-ND2','MT-ND3')
#result <- fetchNode(txtinput=kw, nodetype="gene", searchby="name")
# Query genes by KEGG ids
#kw <- c('hsa:4514','hsa:45356567','hsa:4535')
#result <- fetchNode(txtinput=kw, nodetype="gene", searchby="xref")
# Query proteins by description
#kw <- '6-phosphogluconolactonase'
#result <- fetchNode(txtinput=kw, nodetype="protein", searchby="description", exactmatch=FALSE)
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{fetchRelation}}
}

