% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchRelationship.R
\name{fetchRelationship}
\alias{fetchRelationship}
\title{Query relationships with url}
\usage{
fetchRelationship(url, returnas)
}
\arguments{
\item{url}{a string specifying relationship url.}

\item{returnas}{a string specifying output type. It can be one of dataframe, list, json. Default is dataframe.}
}
\value{
relationship information with the following components:

\code{source, target} = node grinn id

\code{sourcename, targetname} = node name

\code{sourcelabel, targetlabel} = node type

\code{sourcexref, targetxref} = node cross references

\code{type} = relationship type

\code{datasource} = relationship resource

\code{properties} = relationship properties

Return empty list or data frame if error or found nothing.
}
\description{
retrieve relationship information from the database with url. The results include source node, target node, relationship type, dataSource and properties, if applicable.
}
\note{
curlRequest.TRANSACTION is used in other functions to improve speed.
}
\examples{
# Query genes by name
#result <- fetchRelationship("http://localhost:7474/db/data/relationship/0")
#result <- fetchRelationship("http://localhost:7474/db/data/node/0/relationships/out")
#result <- fetchRelationship("http://localhost:7474/db/data/node/0/relationships/in")

}

