% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNetworkOutput.R
\name{formatNetworkOutput}
\alias{formatNetworkOutput}
\title{Format network}
\usage{
formatNetworkOutput(graph, returnas)
}
\arguments{
\item{graph}{list of paths containing network information to be formatted}
}
\value{
list of network information with the following components:

nodes:

\code{id} = node neo4j id

\code{gid} = node grinn id

\code{nodename} = node name

\code{nodelabel} = node type

\code{nodexref} = node cross references

edges:

\code{source, target} = node neo4j id

\code{type} = relationship type

\code{datasource} = relationship resource

\code{properties} = relationship properties

Return empty list if error or found nothing.
}
\description{
format a network or graph output from the database to the data frame format.
}
\note{
maximum no. of paths returned = 30000 paths
}
\examples{
#cypher = "MATCH ptw = (from:Compound)-[:BIOCHEMICAL_REACTION]->(:Compound) WHERE ID(from) = 0 RETURN DISTINCT ptw LIMIT 2"
#nw = curlRequest.TRANSACTION(cypher)
#result = formatNetworkOutput.default(nw, returnas="dataframe")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{fetchNetwork}}, \code{\link{fetchHetNetwork}}, \code{\link{fetchNetworkByGID}}, \code{\link{fetchHetNetworkByGID}}, \url{http://neo4j.com/docs/stable/rest-api-transactional.html}
}

