% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNodeOutput.R
\name{formatNodeOutput}
\alias{formatNodeOutput}
\title{Format resulting nodes}
\usage{
formatNodeOutput(node, returnas)
}
\arguments{
\item{node}{a row of data frame containing node information to be formatted}
}
\value{
node information. Return empty data frame if error or found nothing.
}
\description{
format and return node information for \code{\link{fetchNode}}. Retrieve also node relationships as urls.
}
\examples{
#querystring="UNWIND ['G82952','G82949'] AS x WITH x MATCH (node:Pathway) WHERE node.GID = x RETURN DISTINCT node"
#node = curlRequest(cypher=querystring)
#formatNodeOutput(node[[1]])
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{fetchNode}}
}

