% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphDb.R
\name{getGraphDb}
\alias{getGraphDb}
\title{Get database location}
\usage{
getGraphDb()
}
\value{
url of the database location.
}
\description{
get database location of the current working envornment.
By default the local Neo4j database location is http://localhost:7474/db/data/
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{setGraphDb}}
}

