% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo_test.R
\name{hypo_test}
\alias{hypo_test}
\title{hypothesis test}
\usage{
hypo_test(normalized_data, p, f, factor_name = NULL,
  repeated_factor_name = NULL, confound = NULL)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

