% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aggregated_data.R
\name{load_aggregated_data}
\alias{load_aggregated_data}
\title{Upload aggregated data.}
\usage{
load_aggregated_data(file, ...)
}
\arguments{
\item{file}{an object with a child named datapath, which is the direction of the file input
(object can be input$inputID for shiny fileInput(inputID, ...)).}

\item{...}{Additional arguments for xlsx::read.xlsx2 or read.csv.}
}
\value{
a list of three data frames: "expression"(sample in row), "feature"(compoud in row) and "phenotype"(sample in row).
}
\description{
Upload aggregated data.
Load xlsx or csv file and return a list of expression, feature and phenotype dataset. The standard of xlsx or csv file can be found \link{here}.!!!
}
\details{

}
\examples{
load_aggregated_data(input$inputID,startRow=2)
}
\author{
Sili Fan \email{fansili2013@gmail.com}
}
\seealso{
\code{\link{load_expression_data}}, \code{\link{load_expression_data}}, \code{\link{load_expression_data}}
}

