% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInfile.R
\name{readInfile}
\alias{readInfile}
\title{Read a file}
\usage{
readInfile(file, header, sep, rowname)
}
\description{
read a file overriding read.csv for GUI
}
\examples{
# readInFile("location")
}
\author{
Kwanjeera W \email{kwanich@ucdavis.edu}
}
\seealso{
\code{\link{read.csv}}
}

