% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hypo_test.R
\name{stat_hypo_test}
\alias{stat_hypo_test}
\title{hypothesis test}
\usage{
stat_hypo_test(e, p, f, e_ori, p_ori, e_before, p_before,
  independent_factor_name = NULL, repeated_factor_name = NULL,
  need_power = T, desired_power = NULL, ttestmethod = NULL,
  ttestcorrection = NULL, nonparattestmethod = NULL,
  nonparattestcorrection = NULL, ANOVAmethod = NULL, ANOVAposthoc = NULL,
  nonparaANOVAmethod = NULL, nonparaANOVAposthoc = NULL,
  twowayANOVAmethod = NULL, maineffect1ANOVAmethod = NULL,
  maineffect1posthoc = NULL, simplemaineffect1ANOVAmethod = NULL,
  simplemaineffect1posthoc = NULL, maineffect2ANOVAmethod = NULL,
  maineffect2posthoc = NULL, simplemaineffect2ANOVAmethod = NULL,
  simplemaineffect2posthoc = NULL, nonparatwowayANOVAmethod = NULL,
  nonparamaineffect1ANOVAmethod = NULL, nonparamaineffect1posthoc = NULL,
  nonparasimplemaineffect1ANOVAmethod = NULL,
  nonparasimplemaineffect1posthoc = NULL,
  nonparamaineffect2ANOVAmethod = NULL, nonparamaineffect2posthoc = NULL,
  nonparasimplemaineffect2ANOVAmethod = NULL,
  nonparasimplemaineffect2posthoc = NULL, maineffect1ttestmethod = NULL,
  maineffect1correction = NULL, simplemaineffect1ttestmethod = NULL,
  simplemaineffect1correction = NULL, maineffect2ttestmethod = NULL,
  maineffect2correction = NULL, simplemaineffect2ttestmethod = NULL,
  simplemaineffect2correction = NULL, nonparamaineffect1ttestmethod = NULL,
  nonparamaineffect1correction = NULL,
  nonparasimplemaineffect1ttestmethod = NULL,
  nonparasimplemaineffect1correction = NULL,
  nonparamaineffect2ttestmethod = NULL, nonparamaineffect2correction = NULL,
  nonparasimplemaineffect2ttestmethod = NULL,
  nonparasimplemaineffect2correction = NULL)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

