% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hypo_test.R
\name{stat_hypo_test}
\alias{stat_hypo_test}
\title{hypothesis test}
\usage{
stat_hypo_test(e, p, f, e_ori, p_ori, e_before, p_before,
  independent_factor_name = NULL, repeated_factor_name = NULL,
  need_power = T, desired_power = NULL, onewayANOVAmethod = NULL,
  nonpara_onewayANOVAmethod = NULL, onewayANOVAposthocmethod = NULL,
  nonpara_onewayANOVAposthocmethod = NULL, ttestmethod = NULL,
  ttestFDRmethod = NULL, nonpara_ttestmethod = NULL,
  nonpara_ttestFDRmethod = NULL, twowayANOVAmethod = NULL,
  maineffectANOVAmethod1 = NULL, maineffectANOVAposthocmethod1 = NULL,
  maineffectANOVAmethod2 = NULL, maineffectANOVAposthocmethod2 = NULL,
  simplemaineffectANOVAmethod1 = NULL,
  simplemaineffectANOVAposthocmethod1 = NULL,
  simplemaineffectANOVAmethod2 = NULL,
  simplemaineffectANOVAposthocmethod2 = NULL,
  nonpara_maineffectANOVAmethod1 = NULL,
  nonpara_maineffectANOVAposthocmethod1 = NULL,
  nonpara_maineffectANOVAmethod2 = NULL,
  nonpara_maineffectANOVAposthocmethod2 = NULL,
  nonpara_simplemaineffectANOVAmethod1 = NULL,
  nonpara_simplemaineffectANOVAposthocmethod1 = NULL,
  nonpara_simplemaineffectANOVAmethod2 = NULL,
  nonpara_simplemaineffectANOVAposthocmethod2 = NULL,
  maineffectttestmethod1 = NULL, maineffectttestposthocmethod1 = NULL,
  maineffectttestmethod2 = NULL, maineffectttestposthocmethod2 = NULL,
  simplemaineffectttestmethod1 = NULL,
  simplemaineffectttestposthocmethod1 = NULL,
  simplemaineffectttestmethod2 = NULL,
  simplemaineffectttestposthocmethod2 = NULL,
  nonpara_maineffectttestmethod1 = NULL,
  nonpara_maineffectttestposthocmethod1 = NULL,
  nonpara_maineffectttestmethod2 = NULL,
  nonpara_maineffectttestposthocmethod2 = NULL,
  nonpara_simplemaineffectttestmethod1 = NULL,
  nonpara_simplemaineffectttestposthocmethod1 = NULL,
  nonpara_simplemaineffectttestmethod2 = NULL,
  nonpara_simplemaineffectttestposthocmethod2 = NULL,
  onewayrepeatedANOVAmethod = "rANOVA",
  onewayrepeatedANOVAposthocmethod = "paired+bonf", onewaySpher_Corr = NULL,
  nonpara_onewayrepeatedANOVAmethod = "rANOVA",
  nonpara_onewayrepeatedANOVAposthocmethod = "nonpara_paired+bonf",
  pairedttestmethod = "paired t test", pairedttestFDRmethod = NULL,
  nonpara_pairedttestmethod = "Wil test",
  nonpara_pairedttestFDRmethod = NULL, mainSpher_Corr1 = NULL,
  mainSpher_Corr2 = NULL, simplemainSpher_Corr1 = NULL,
  simplemainSpher_Corr2 = NULL,
  simplemaineffectpairedttestposthocmethod1 = NULL,
  simplemaineffectpairedttestposthocmethod2 = NULL,
  nonpara_simplemaineffectpairedttestposthocmethod1 = NULL,
  nonpara_simplemaineffectpairedttestposthocmethod2 = NULL,
  maineffectpairedttestposthocmethod1 = NULL,
  nonpara_maineffectpairedttestposthocmethod1 = NULL,
  maineffectpairedttestposthocmethod2 = NULL,
  nonpara_maineffectpairedttestposthocmethod2 = NULL)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

