% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hypo_test.R
\name{stat_hypo_test}
\alias{stat_hypo_test}
\title{hypothesis test}
\usage{
stat_hypo_test(e, p, f, e_ori, p_ori, e_before, p_before,
  independent_factor_name = NULL, repeated_factor_name = NULL,
  need_power = T, desired_power = NULL, ttestmethod = NULL,
  ttestcorrection = NULL, nonparattestmethod = NULL,
  nonparattestcorrection = NULL, ANOVAmethod = NULL, ANOVAposthoc = NULL,
  nonparaANOVAmethod = NULL, nonparaANOVAposthoc = NULL)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

