% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hypo_test.R
\name{stat_hypo_test}
\alias{stat_hypo_test}
\title{hypothesis test}
\usage{
stat_hypo_test(e, p, f, e_ori, p_ori, e_before, p_before,
  independent_factor_name = NULL, repeated_factor_name = NULL,
  need_power = T, desired_power = NULL, ttestmethod = NULL,
  ttestcorrection = NULL, nonparattestmethod = NULL,
  nonparattestcorrection = NULL, ANOVAmethod = NULL, ANOVAposthoc = NULL,
  nonparaANOVAmethod = NULL, nonparaANOVAposthoc = NULL,
  twowayANOVAmethod = "two way ANOVA", maineffect1ANOVAmethod = "Welch",
  maineffect1ANOVAposthoc = "games.howell",
  simplemaineffect1ANOVAmethod = "Welch",
  simplemaineffect1ANOVAposthoc = "games.howell",
  maineffect1ttestmethod = "Welch", maineffect1ttestcorrection = "fdr",
  simplemaineffect1ttestmethod = "Welch",
  simplemaineffect1ttestcorrection = "fdr", maineffect2ANOVAmethod = NULL,
  maineffect2ANOVAposthoc = NULL, simplemaineffect2ANOVAmethod = NULL,
  simplemaineffect2ANOVAposthoc = NULL, maineffect2ttestmethod = "Welch",
  maineffect2ttestcorrection = "fdr",
  simplemaineffect2ttestmethod = "Welch",
  simplemaineffect2ttestcorrection = "none",
  nonparatwowayANOVAmethod = NULL, nonparamaineffect1ANOVAmethod = "htest",
  nonparamaineffect1ANOVAposthoc = "utest",
  nonparasimplemaineffect1ANOVAmethod = "htest",
  nonparasimplemaineffect1ANOVAposthoc = "utest",
  nonparamaineffect1ttestmethod = "utest",
  nonparamaineffect1ttestcorrection = "fdr",
  nonparasimplemaineffect1ttestmethod = "uteste",
  nonparasimplemaineffect1ttestcorrection = "fdr",
  nonparamaineffect2ANOVAmethod = NULL,
  nonparamaineffect2ANOVAposthoc = NULL,
  nonparasimplemaineffect2ANOVAmethod = NULL,
  nonparasimplemaineffect2ANOVAposthoc = NULL,
  nonparamaineffect2ttestmethod = "utest",
  nonparamaineffect2ttestcorrection = "fdr",
  nonparasimplemaineffect2ttestmethod = "utest",
  nonparasimplemaineffect2ttestcorrection = "none",
  pairedttestmethod = NULL, pairedttestcorrection = NULL,
  nonparapairedttestmethod = NULL, nonparapairedttestcorrection = NULL,
  pairedANOVAmethod = NULL, pairedANOVAadjust = NULL,
  pairedANOVAposthoc = NULL, nonparapairedANOVAmethod = NULL,
  nonparapairedANOVAposthoc = NULL, twowaypairedANOVAmethod = NULL,
  twowaypairedANOVAadjust = NULL, maineffect1pairedANOVAmethod = NULL,
  maineffect1pairedANOVAadjust = NULL, maineffect1pairedANOVAposthoc = NULL,
  simplemaineffect1pairedANOVAmethod = NULL,
  simplemaineffect1pairedANOVAadjust = NULL,
  simplemaineffect1pairedANOVAposthoc = NULL,
  maineffect1pairedttestmethod = "Welch",
  maineffect1pairedttestcorrection = "fdr",
  simplemaineffect1pairedttestmethod = "utest",
  simplemaineffect1pairedttestcorrection = "fdr",
  maineffect2pairedANOVAmethod = NULL, maineffect2pairedANOVAadjust = NULL,
  maineffect2pairedANOVAposthoc = NULL,
  simplemaineffect2pairedANOVAmethod = NULL,
  simplemaineffect2pairedANOVAadjust = NULL,
  simplemaineffect2pairedANOVAposthoc = NULL,
  maineffect2pairedttestmethod = "paired t test",
  maineffect2pairedttestcorrection = "fdr",
  simplemaineffect2pairedttestmethod = "paired t test",
  simplemaineffect2pairedttestcorrection = "fdr",
  nonparatwowaypairedANOVAmethod = NULL,
  nonparamaineffect1pairedANOVAmethod = NULL,
  nonparamaineffect1pairedANOVAposthoc = NULL,
  nonparasimplemaineffect1pairedANOVAmethod = NULL,
  nonparasimplemaineffect1pairedANOVAposthoc = NULL,
  nonparamaineffect1pairedttestmethod = "paired t test",
  nonparamaineffect1pairedttestcorrection = "fdr",
  nonparasimplemaineffect1pairedttestmethod = "paired utest",
  nonparasimplemaineffect1pairedttestcorrection = "fdr",
  nonparamaineffect2pairedANOVAmethod = NULL,
  nonparamaineffect2pairedANOVAposthoc = NULL,
  nonparasimplemaineffect2pairedANOVAmethod = NULL,
  nonparasimplemaineffect2pairedANOVAposthoc = NULL,
  nonparamaineffect2pairedttestmethod = "paired utest",
  nonparamaineffect2pairedttestcorrection = "fdr",
  nonparasimplemaineffect2pairedttestmethod = "paired utest",
  nonparasimplemaineffect2pairedttestcorrection = "fdr",
  mixedANOVA = "mixed anova", mixedANOVAadjust = "GG",
  nonparamixedANOVA = "nonpara mixed anova", bootstrap = FALSE,
  bootstrap_num = 500)
}
\description{
hypothesis test
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

