% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_norm.R
\name{stat_norm}
\alias{stat_norm}
\title{normalization}
\usage{
stat_norm(e, f, p, sample_index, mTICdid, Loessdid, medFCdid, BatchMediandid,
  mTIC, Loess, medFC, BatchMedian, sample_normalization = NULL,
  data_transformation = NULL, data_scaling = NULL, log_para,
  independent_factor_name_log, power_para, independent_factor_name_power)
}
\description{
normalization
}
\details{

}
\author{
Sili Fan \email{fansili2013@gmail.com}
}

